Analyzing summarized data

Apply a statistical test to the data.

Usage

analyze(
  dataSet,
  method = "t-test",
  ref = NULL,
  adjust.method = "none",
  paired = FALSE,
  pool.sd = FALSE,
  center = TRUE,
  scale = TRUE
)

Arguments

dataSet

The 2d data set of data.

method

A string (default = "t-test") specifying which statistical analysis to use:

"t-test": Student's t-test.

"mod.t-test": Empirical Bayes moderated t-test (Smyth 2004) .

"wilcox-test": Wilcoxon test.

"MA": Output to plot an MA plot.

"PCA": Principal components analysis (Pearson 1901; Hotelling 1933) .

ref

A string (default = NULL) specifying the reference condition for comparison when method = "*-test" or method = "MA". If NULL, all pairwise comparisons are performed.

adjust.method

A string (default = "none") specifying the correction method for p-value adjustment when method = "*-test":

"BH" or its alias "fdr": Benjamini and Hochberg (1995) .

"BY": Benjamini and Yekutieli (2001) .

"bonferroni": Bonferroni (1936) .

"hochberg": Hochberg (1988) .

"holm": Holm (1979) .

"hommel": Hommel (1988) .

"none": None

See p.adjust for more details.

paired

A boolean (default = FALSE) specifying whether or not to perform a paired test when method = "t-test" or method = "wilcox-test".

pool.sd

A boolean (default = FALSE) specifying whether or not to use a pooled standard deviation when method = "t-test".

center

A boolean (default = TRUE) indicating whether the variables should be shifted to be zero centered when method = "PCA".

scale

A boolean (default = TRUE) indicating whether the variables should be scaled to have unit variance before the analysis takes place when method = "PCA".

Value

"t-test", "mod.t-test", "wilcox-test": A list comprising data frames for each comparison, with each data frame containing the means of the two compared conditions for each protein, the difference in means, and the p-values. Additionally, a separate data frame called "total" summarizes the results of multiple comparisons.

"MA": A list comprising data frames for each comparison, with each data frame containing the means of the two compared conditions for each protein, as well as the average and difference in means.

"PCA": A list containing the standard deviations of the principal components sdev, the matrix of variable loadings rotation, the centering used center, the scaling used scale, and the principal component scores x.

References

Benjamini Y, Hochberg Y (1995). “Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing.” Journal of the Royal Statistical Society: Series B (Methodological), 57(1), 289–300. doi:10.1111/j.2517-6161.1995.tb02031.x .

Benjamini Y, Yekutieli D (2001). “The Control of the False Discovery Rate in Multiple Testing under Dependency.” The Annals of Statistics, 29(4), 1165–1188. doi:10.1214/aos/1013699998 .

Bonferroni CE (1936). “Teoria Statistica Delle Classi e Calcolo Delle Probabilità.” Pubblicazioni del R Istituto Superiore di Scienze Economiche e Commerciali di Firenze, 8, 3–62.

Hochberg Y (1988). “A Sharper Bonferroni Procedure for Multiple Tests of Significance.” Biometrika, 75(4), 800–802. doi:10.1093/biomet/75.4.800 .

Holm S (1979). “A Simple Sequentially Rejective Multiple Test Procedure.” Scandinavian Journal of Statistics, 6(2), 65–70. https://www.jstor.org/stable/4615733.

Hommel G (1988). “A Stagewise Rejective Multiple Test Procedure Based on a Modified Bonferroni Test.” Biometrika, 75(2), 383–386. doi:10.1093/biomet/75.2.383 .

Hotelling H (1933). “Analysis of a Complex of Statistical Variables into Principal Components.” Journal of Educational Psychology, 24(6), 417–441. doi:10.1037/h0071325 .

Pearson K (1901). “On Lines and Planes of Closest Fit to Systems of Points in Space.” The London, Edinburgh, and Dublin Philosophical Magazine and Journal of Science, 2(11), 559–572. doi:10.1080/14786440109462720 .

Smyth GK (2004). “Linear Models and Empirical Bayes Methods for Assessing Differential Expression in Microarray Experiments.” Statistical Applications in Genetics and Molecular Biology, 3(1). doi:10.2202/1544-6115.1027 .