Filter proteins by gene, accession or description

Filter the preprocessed dataset by gene, accession, or description.

Usage

filterProtein(
  dataSet,
  proteinInformation = "preprocess_protein_information.csv",
  text = c(),
  by = c("PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions", "PG.ProteinNames",
    "ProteinDescriptions", "AccessionNumber", "AlternateID"),
  removeList = TRUE,
  saveRm = TRUE
)

Arguments

dataSet

The 2d data set of experimental values.

proteinInformation

The name of the .csv file containing protein information data (including the path to the file, if needed). The file should include the following columns:

For Spectronaut: "PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions", and "PG.ProteinNames".

For Scaffold: "ProteinDescriptions", "AccessionNumber", and "AlternateID".

This file is automatically generated by the function preprocessing or preprocessing_scaffold.

text

A character vector of text used as the key for selecting or removing.

by

A character string specifying the information to which the text filter is applied, with allowable options:

For Spectronaut: "PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions", and "PG.ProteinNames".

For Scaffold: "ProteinDescriptions", "AccessionNumber", and "AlternateID".

removeList

A boolean (default = TRUE) specifying whether the list of proteins should be removed or selected.

TRUE: Remove the list of proteins from the data set.

FALSE: Remove all proteins not in the list from the data set.

saveRm

A boolean (default = TRUE) specifying whether to save removed data to current working directory. This option only works when removeList = TRUE.

Value

A filtered 2d dataframe.

Details

The function is an extension of the function preprocessing or preprocessing_scaffold that allows for filtering proteins based on additional information.