Imputation of raw data signals

Apply an imputation method to the data set.

Usage

impute(
  dataSet,
  imputeType = "LocalMinVal",
  reqPercentPresent = 0.51,
  k = 10,
  rowmax = 0.5,
  colmax = 0.8,
  maxp = 1500,
  rng.seed = 362436069,
  reportImputing = FALSE
)

Arguments

dataSet

The 2d data set of experimental values.

imputeType

A string (default = "LocalMinVal") specifying which imputation method to use:

"LocalMinVal": replace missing values with the lowest value from the protein by condition combination.

"GlobalMinVal": replaces missing values with the lowest value found within the entire data set.

"knn": replace missing values using the k-nearest neighbors algorithm.

"seq-knn": replace missing values using the sequential k-nearest neighbors algorithm.

reqPercentPresent

An numeric value (default = 0.51) specifying the required percent of values that must be present in a given protein by condition combination for values to be imputed when imputeType = "LocalMinVal".

k

An integer (default = 10) indicating the number of neighbors to be used in the imputation when imputeType = "knn" or imputeType = "seq-knn".

rowmax

A numeric value (default = 0.5) specifying the maximum percent missing data allowed in any row when imputeType = "knn". For any rows with more than rowmax*100% missing are imputed using the overall mean per sample.

colmax

A numeric value (default = 0.8) specifying the maximum percent missing data allowed in any column when imputeType = "knn". If any column has more than colmax*100% missing data, the program halts and reports an error.

maxp

An integer (default = 1500) indicating the largest block of proteins imputed using the k-nearest neighbors algorithm when imputeType = "knn". Larger blocks are divided by two-means clustering (recursively) prior to imputation.

rng.seed

An integer (default = 362436069) specifying the seed used for the random number generator for reproducibility when imputeType = "knn".

reportImputing

A boolean (default = FALSE) specifying whether to provide a shadow data frame with imputed data labels. Alters the return structure.

Value

If reportImputing = FALSE, the function returns the imputed 2d dataframe.

If reportImputing = TRUE, the function returns a list of the imputed 2d dataframe and a shadow matrix showing which proteins by replicate were imputed.

References

Troyanskaya, O., Cantor, M., Sherlock, G., Brown, P., Hastie, T., Tibshirani, R., Botstein, D. and Altman, R. B. (2001). Missing Value Estimation Methods for DNA Microarrays. Bioinformatics, 17(6), 520--525.

Kim, KY., Kim, BJ. and Yi, GS. (2004). Reuse of Imputed Data in Microarray Analysis Increases Imputation Efficiency. BMC Bioinformatics, 5, 160.