Normalization of preprocessed data

Apply a specified type of normalization to a data set.

Usage

normalize(dataSet, normalizeType = "quant")

Arguments

dataSet

The 2d data set of experimental values.

normalizeType

A string (default = "quant") specifying which type of normalization to apply:

"quant": Quantile (Bolstad et al. 2003)

"median": Protein-wise Median

"mean": Protein-wise Mean

"none": None

Value

A normalized 2d dataframe.

Details

Quantile normalization is generally recommended. Mean and median normalization are going to be included as popular previous methods. No normalization is not recommended. Boxplots are also generated for before and after the normalization to give a visual indicator of the changes.

References

Bolstad BM, Irizarry RA, Astrand M, Speed TP (2003). “A Comparison of Normalization Methods for High Density Oligonucleotide Array Data Based on Variance and Bias.” Bioinformatics, 19(2), 185–193. doi:10.1093/bioinformatics/19.2.185 .