Compiling data on a single protein from each step in the process

Summarize the steps performed on the data for one protein.

Usage

pullProteinPath(proteinName, dataSetList)

Arguments

proteinName

A string identifying the protein of interest.

dataSetList

A list of data frames, the order dictates the order of presentation.

Value

A 2d dataframe, with the protein data at each step present in the dataSetList.

Details

proteinName must match the labels in the data sets exactly.