Generating visualizations for MS Data

Create specific graphics to illustrate the results of the data analysis function.

visualize(
  dataSet,
  graphType = "volcano",
  pkg = "pheatmap",
  cluster_cols = TRUE,
  cluster_rows = FALSE,
  show_colnames = TRUE,
  show_rownames = TRUE,
  conditions = c(),
  transformType = NULL,
  addlabels = TRUE,
  choice = "variance",
  ncp = 10,
  addEllipses = TRUE,
  ellipse.level = 0.95,
  label = "all",
  fileName,
  P.thres = 0.05,
  logF.thres = 0.6
)

Arguments

dataSet

A 2D data frame that corresponds to the output from the function analyze() for the same name.

graphType

A string indicating the graph type. Current options are:

"heatmap"

"MA"

"normalize"

"PCA_biplot"

"PCA_ind"

"PCA_scree"

"PCA_var"

"t-test"

"venn"

"volcano"

pkg

A string specifying the source package used to plot the heatmap. Two options: "pheatmap" and "ggplot2". This argument only works when graphType = "heatmap".

cluster_cols

A boolean (default = TRUE) determining if rows should be clustered or hclust object. This argument only works when graphType = "heatmap" and pkg = "pheatmap".

cluster_rows

A boolean (default = FALSE) determining if columns should be clustered or hclust object. This argument only works when graphType = "heatmap" and pkg = "pheatmap".

show_colnames

A boolean (default = TRUE) specifying if column names are be shown. This argument only works when graphType = "heatmap" and pkg = "pheatmap".

show_rownames

A boolean (default = TRUE) specifying if row names are be shown. This argument only works when graphType = "heatmap" and pkg = "pheatmap".

conditions

A string specifying which two conditions to compare when graphType = "MA". This argument only works when there are more than two conditions in dataSet.

transformType

A string used to label the title of the transformation type for the transformed MA plot when graphType = "MA".

addlabels

A boolean (default = TRUE) specifying whether the elements are labeled.

For graphType = "PCA_scree", it specifies whether labels are added at the top of bars or points to show the information retained by each dimension.

For graphType = "PCA_ind", it specifies whether the active individuals to be labeled.

For graphType = "PCA_var", it specifies whether the active variables to be labeled.

choice

A text (default = "variance") specifying the PCA data to be plotted the scree plot when graphType = "PCA_scree". Allowed values are "variance" or "eigenvalue".

ncp

A numeric value (default = 10) specifying the number of dimensions to be shown when graphType = "PCA_scree".

addEllipses

A boolean (default = TRUE) specifying whether to draw ellipses around the individuals when graphType = "PCA_ind" or graphType = "PCA_biplot".

ellipse.level

A numeric value (default = 0.95) specifying the size of the concentration ellipse in normal probability when graphType = "PCA_ind" or graphType = "PCA_biplot".

label

A text (default = "all") specifying the elements to be labelled when graphType = "PCA_biplot". Allowed values:

"all": label both active individuals and active variables.

"ind": label only active individuals.

"var": label only active variables.

"none": no labels.

fileName

Filename specifying the name for Venn image output, or if NULL, it returns the grid object itself.

P.thres

THe threshold value of P-value (default = 0.05) used to plot the horizontal line (-log10(P.thres)) on the volcano plot when graphType = "volcano".

logF.thres

The absolute threshold value of log2(fold change) (default = 0.6) used to plot the two vertical lines (-logF.thres and logF.thres) on the volcano plot when graphType = "volcano".

Value

An object of class ggplot.

Details

The function visualize() is designed to work directly with output from the function analyze(). Please be sure that the arguments graphType and testType match.