Analyzing summarized data

Apply a statistical test to the data.

Usage

analyze(dataSet, conditions, testType = "t-test")

Arguments

dataSet

The 2d data set of data.

conditions

A string specifying which two conditions to compare. This argument only works when there are more than two conditions in dataSet.

testType

A string (default = "t-test") specifying which statistical test to use:

"t-test": unequal variance t-test.

"mod.t-test": moderated t-test (Smyth 2004) .

"MA": output to plot an MA plot.

Value

A 2d dataframe includes the following information:

"t-test" or "mod.t-test": The differences in means and P-values for each protein between the two conditions. Note that the differences are calculated by subtracting the mean of the second condition from the mean of the first condition (Condition 1 - Condition 2).

"MA": Protein-wise averages within each condition.

References

Smyth GK (2004). “Linear Models and Empirical Bayes Methods for Assessing Differential Expression in Microarray Experiments.” Statistical Applications in Genetics and Molecular Biology, 3(1). doi:10.2202/1544-6115.1027 .