Compiling data on a single protein from each step in the process

Summarize the steps performed on the data for one protein.

Usage

pullProteinPath(
  listName = NULL,
  regexName = NULL,
  dataSetList,
  by = "PG.ProteinNames",
  proteinInformation = "preprocess_protein_information.csv"
)

Arguments

listName

A character vector identifying the proteins of interest.

regexName

A character vector specifying the proteins for regular expression pattern matching.

dataSetList

A list of data frames, the order dictates the order of presentation.

by

A character string specifying the information to which the listName and/or regexName are applied, with allowable options:

For Spectronaut: "PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions", and "PG.ProteinNames".

For Scaffold: "ProteinDescriptions", "AccessionNumber", and "AlternateID".

proteinInformation

The name of the .csv file containing protein information data (including the path to the file, if needed). The file should include the following columns:

For Spectronaut: "PG.Genes", "PG.ProteinAccessions", "PG.ProteinDescriptions", and "PG.ProteinNames".

For Scaffold: "ProteinDescriptions", "AccessionNumber", and "AlternateID".

This file is automatically generated by the function preprocessing or preprocessing_scaffold.

Value

A 2d dataframe, with the protein data at each step present in the dataSetList.